Adds support for two more formatting options for German.
The behaviour of %e is changed to output ordinal numbers using numerals, as was specified in the docblock. %E is newly intoduced to support the former behaviour of %e (ordinal numbers as words). %C is newly introduced to output the seasons in casus genitive with leading article (in English you can just write "of" into the format string. In German however the season names have different genus, therefore it is sometimes "der" and sometimes "des" and in the latter case also an "s" is added as a suffix to the season name). Also fixes issues in the German formatter that made the unit tests fail, after the English formatter was changed.